requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Stock Data & Analysis
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
pandas-ta==0.3.14b

# AI & ML - FIX: Add scikit-learn
google-generativeai==0.3.2
scikit-learn==1.3.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
